feat(Cleaning): moves links atoms to a dedicated folder and stories
