+++ src/components/atoms/Link/ExternalLink.ts
@@ -4,6 +4,10 @@ import Anchor from './Anchor';
 const ExternalLink = styled(Anchor).attrs({
   target: '_blank',
   rel: 'noopener noreferrer'
-})``;
+})`
+  & > svg {
+    fill: ${props => props.theme.grey};
+  }
+`;
 
 export default ExternalLink;
