+++ src/components/atoms/Link/index.ts
@@ -1,0 +1,3 @@
+export { default as Anchor } from './Anchor';
+export { default as Link } from './Link';
+export { default as ExternalLink } from './ExternalLink';
