+++ src/components/atoms/icons/Link.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export default () => (
+  <svg xmlns="http://www.w3.org/2000/svg" width="16.53" height="16.533">
+    <path d="M9.276 12.045h-.11a4.637 4.637 0 01-1.763-.344.276.276 0 00-.3.061l-2.08 2.083a1.654 1.654 0 11-2.339-2.339l3.518-3.515a1.653 1.653 0 012.336 0 1.13 1.13 0 001.548 0 1.091 1.091 0 000-1.548 3.75 3.75 0 00-.647-.521 3.858 3.858 0 00-4.807.521l-3.5 3.507a3.857 3.857 0 005.452 5.458l2.893-2.893a.276.276 0 00-.2-.471z" />
+    <path d="M15.4 1.13a3.857 3.857 0 00-5.455 0l-2.89 2.887a.277.277 0 00.2.471h.1a4.628 4.628 0 011.76.344.275.275 0 00.3-.061L11.494 2.7a1.654 1.654 0 112.339 2.339L11.249 7.62l-.022.025-.9.9a1.653 1.653 0 01-2.336 0 1.13 1.13 0 00-1.548 0 1.1 1.1 0 000 1.554 3.824 3.824 0 001.1.771c.058.028.116.05.174.074s.118.044.176.066.118.041.176.058l.163.044c.11.028.22.05.333.069a3.834 3.834 0 00.411.039h.209l.165-.019c.061 0 .124-.017.2-.017h.094l.19-.028.088-.017.16-.033h.03a3.856 3.856 0 001.788-1.014l3.507-3.507A3.857 3.857 0 0015.4 1.13z" />
+  </svg>
+);
