+++ src/components/atoms/index.ts
@@ -16,9 +16,9 @@ export { default as Contributor } from './Contributor';
 export { default as Time } from './Time';
 export { default as Timer } from './Timer';
 export { default as List } from './List';
-export { default as ExternalLink } from './ExternalLink';
-export { default as Link } from './Link';
-export { default as Anchor } from './Anchor';
+export { default as ExternalLink } from './Link/ExternalLink';
+export { default as Link } from './Link/Link';
+export { default as Anchor } from './Link/Anchor';
 export { default as Title } from './Title';
 
 export { default as Box } from './Box/Box';
